+++ TurismoArgentina_backend/src/main/java/com/mgbt/turismoargentina_backend/controllers/LocationController.java
@@ -178,7 +178,7 @@ public class LocationController {
             content = { @Content(mediaType = "application/json", schema = @Schema(implementation = Long.class)) })
     public ResponseEntity<?> getCount(Locale locale) {
         try {
-            Long count = this.locationService.getCount();
+            int count = this.locationService.getCount();
             return new ResponseEntity<>(count, HttpStatus.OK);
         } catch (DataAccessException ex) {
             return this.exceptionService.throwDataAccessException(ex, locale);
