+++ TurismoArgentina_backend/src/main/java/com/mgbt/turismoargentina_backend/model/repositories/IProvinceRepository.java
@@ -31,5 +31,5 @@ public interface IProvinceRepository extends JpaRepository<Province, Long> {
 
     @Query(value = "SELECT COUNT(*) FROM turismo_argentina.provinces p " +
             "WHERE p.deletion_date IS NULL", nativeQuery = true)
-    Long findCount();
+    int findCount();
 }
