+++ TurismoArgentina_backend/src/main/java/com/mgbt/turismoargentina_backend/model/services/IActivityService.java
@@ -17,6 +17,6 @@ public interface IActivityService {
     List<Activity> getFiveRandom();
     Page<Activity> getByLocationName(Pageable pageable, String locationName);
     List<Activity> getByLocationId(Long idLocation);
-    Long getCount();
+    int getCount();
     String getPreviousImage(Activity activity, Integer imageNumber, String fileName);
 }
