+++ TurismoArgentina_backend/src/main/java/com/mgbt/turismoargentina_backend/model/services/ILocationService.java
@@ -19,5 +19,5 @@ public interface ILocationService {
     List<String> getAllLocationNames();
     Page<Location> getByProvinceName(Pageable pageable, String provinceName);
     List<Location> getByProvinceId(Long idProvince);
-    Long getCount();
+    int getCount();
 }
