+++ TurismoArgentina_backend/src/main/java/com/mgbt/turismoargentina_backend/model/services/IProvinceService.java
@@ -17,5 +17,5 @@ public interface IProvinceService {
     Province findByName(String name);
     List<Province> getThreeRandom();
     List<String> getAllProvinceNames();
-    Long getCount();
+    int getCount();
 }
